+++ resources/views/user/decisions/index.blade.php
