+++ resources/views/user/decisions/show.blade.php
